pyproject.toml
[project]
name = "realtime-kline-dashboard"
version = "0.1.0"
description = "Pulling data from binance websocket for live data visualization."
dependencies = [
    "lightweight-charts>=2.1",
    "requests>=2.32.5",
    "websocket-client>=1.8.0",
    "pybit>=5.12.0",
    "google-genai>=1.45.0",
    "anthropic>=0.71.0",
    "Markdown>=3.9"

]
UV_ENV_FILE=".env"
